requirements.txt
telethon
lyricsgenius
pythondialog
cryptg
Pillow
future-fstrings
